Updated 'destination' and 'source' properties.
Connecting an AudioPlayer or AudioOsc source can now be done without the 'sourceName.buffer'.
Just link to the 'sourceName' and the buffer will be selected by default.  

The same goes for the goes for the 'destination' property on an AudioEffect, AudioConvolver, AudioDistortion, AudioGain, AudioPlayer.play and AudioOsc.play.  
Set the destination to the variable you want to connect it to.